+++ NEWS.md
@@ -1,8 +1,14 @@
-# fmpcloudr 0.1.1
+# fmpcloudr 0.1.2
 
 ## Release Notes and News
 
+### fmpcloudr 0.1.2 1/20/2021
+Converted fmpc_symbols_sp500 to fmpc_symbols_index
+to allow for the pulling of S&P 500, Nasdaq, and Dow Jones.
+
 ### fmpcloudr 0.1.1 10/24/2020
+Approved by CRAN: 2020-11-03 with commit 3089cbc.
+
 Incorporates CRAN requested changes on URLs in description,
 exporting fmpc_security_news function, and using on.exit(),
 in stock screener function to reset options.